+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ##### Tuesday, June 18, 2019
-- Added gh-pages branch
+- Added *gh-pages* branch
+- Added git submodules for *pelican-themes* and *pelican-plugins*
+- Added *themes.py* to hold names of all the themes from *pelican-themes* in an list called Themes
+